requirements.txt
snowflake-connector-python
langchain-core
langchain_openai
fastapi
uvicorn
openai
requests
streamlit
pandas
python-dotenv
selenium
boto3
sentence-transformers
numpy
pinecone
git+https://github.com/brandonstarxel/chunking_evaluation.git
tavily-python
langgraph
ipython
matplotlib